Prevent possible memory leak in TabletServiceImpl
When we allocate a WriteTransactionState in TabletServiceImpl::Write()
we would do so on the heap, but without assigning it to a smart pointer.
Then, right after that, we check that can update the clock and return an
error to the user if we can't. If we returned an error WriteTransctionState
would leak.

This changes the allocation statement to be made closer to where it is
used, after the clock is updated. This also makes it use a smart pointer
(unique_ptr) there and wherever else required. Finally this changes
AlterSchemaTransactionState to use the same smart pointer, for consistency.

Change-Id: I97181d59ea7e2d34067ad1c72edda2d6fc7360d5
Reviewed-on: http://gerrit.cloudera.org:8080/3082
Tested-by: David Ribeiro Alves <david.alves@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>