Prevent flakiness in ClientTest::TestFailedDnsResolution()
We're currently observing flakiness in this test because the timeout
is set so low that we might get a Status::TimeOut() before we can get
the dns resolution error that we're expecting.

This just sets the timeout to be a little higher, to 500 ms.
I've spun the test for 100 cycles and can't observe the error anymore.

Change-Id: I20a00d3a08e3576cbbf7070527b98081cd1b6ae6
Reviewed-on: http://gerrit.cloudera.org:8080/3083
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>