Remove the use of boost::lock_guard in oid_generator.cc
When using the OidGenerator class in the client we would get:

/usr/bin/ld: /usr/lib/x86_64-linux-gnu/libboost_system.a(error_code.o): relocation R_X86_64_32 against `.rodata.str1.1' can not be used when making a shared object; recompile with -fPIC

That is, building and linking the client would now require libboost_system.a to
be compiled with -fPIC, which it isn't in some distros.

Adar suggested that if we removed the use of boost::lock_guard in
oid_generator.cc we might be able to forego the dependency on libboost_system
by the client altogether, at least for now. This turned out to be true, thus
this patch makes that change.

Change-Id: I224bda0faaf57ed6d6cce0b223d3484317b83342
Reviewed-on: http://gerrit.cloudera.org:8080/3090
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins