KUDU-1446 Consider data types in predicate evaluation order
This patch enhances SelectivityRank() in column_predicate.cc to account for
column type. Predicate type dominates column type for the sort, and a predicate
on a column of a small type will sort before a predicate on a column of larger
(or variable-size) type.

It also adds a small test for selectivity comparison.

Change-Id: I9c3388876035cb494fdb2c0e1f17698f089fdf72
Reviewed-on: http://gerrit.cloudera.org:8080/3298
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>