catalog_manager: move VLOG(1) messages above Async RPC calls to avoid crash
Several functions in this file used a pattern of sending an Async RPC, followed
by logging that the RPC had been sent. However, the callback for the async RPC
could delete the 'this' instance, which would cause the log message to crash.

This only showed up as flakiness in Java unit tests, because they set --v=1
and these messages were VLOG(1).

Change-Id: Ic6beaf940fe42a0c7a3aa5877b7ef65136af210b
Reviewed-on: http://gerrit.cloudera.org:8080/3181
Reviewed-by: Jean-Daniel Cryans
Tested-by: Kudu Jenkins