KUDU-1467. Fix incorrect bootstrap replay issue with UPSERT
This fixes an issue with UPSERT that was causing fuzz-itest to be ~5% flaky.
The issue turned out to be essentially the same as KUDU-1341 (fixed in
1ff209e85b4c2cc4beda7560d328b7ed05e008d2) except pertaining to UPSERT
operations which acted as mutations, whereas the original bug was
for direct UPDATE operations.

In addition to the existing fuzz test which exposes this bug, this patch
includes a specific minimal fuzz sequence which triggered the bug
deterministically.

To verify that the bug is fixed, I looped fuzz-itest 500 times:
http://dist-test.cloudera.org/job?job_id=todd.1464152165.32007

Change-Id: Id443e171bc80833737486a96a3fd6fcce7a2363b
Reviewed-on: http://gerrit.cloudera.org:8080/3208
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans
Reviewed-by: David Ribeiro Alves <david.alves@cloudera.com>
(cherry picked from commit f9aa4ee37c07dd03af792fa99f138b9b5ca8a557)
Reviewed-on: http://gerrit.cloudera.org:8080/3216