Specify guaranteed semantics of GetTableLocations RPC
This commit specifies additional semantics guaranteed by the master's
GetTableLocations RPC. Technically this is a breaking change as some of the
existing behavior is changed, but in practice no known published client version
is affected. These guarantees are necessary in order to implement a meta cache
in the presence of non-covering partitions.

Change-Id: Ibc2b5b647e33c0ee8fc052192870f814eff8c178
Reviewed-on: http://gerrit.cloudera.org:8080/3240
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>