Make 'num_attempts' in RequestIdPB required
We need to differentiate between attempts, on the server, so that we
can handle dangling handlers when multiple attempts at the same RPC
are happening at the same time (e.g. a client-originated write rpc
and the same rpc originating from another replica).

Differentiating which attempt is executing on the server side
requires distinguishing between different attempts. Using 'num_attempts'
is a natural way to do so.

Change-Id: Idd8f780826438278993554546edfae90bf1a39df
Reviewed-on: http://gerrit.cloudera.org:8080/3408
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins