[java-client] use tablet cache for locateTablet calls
This simplifies and improves the performance of AsyncKuduClient::locateTable by
using the client's existing tablet cache. This change is motivated by an
upcoming fix to the locate table logic, which will be easier without the code
duplication.

The public KuduTable::locateTable methods have been deprecated since they rely
on callers having knowledge of the private internals of Kudu's partitioning
implementation. Additionally, the methods have been changed to take an exclusive
end partition key instead of an inclusive end partition key. Whether the key was
inclusive or exclusive was previously undocumented, and the only known caller
(the internal ScanToken implementation) was using it as exclusive. Users of
these methods are encouraged to use the ScanToken API instead, since it's easier
to use and does not leak internal Kudu implementation details.

Change-Id: I78b5d400778547a9ee090111663677901dbadd98
Reviewed-on: http://gerrit.cloudera.org:8080/3386
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans