Update Java client for new master GetTableLocations semantics
The Java client was relying on the master returning an empty tablet locations
response when the table was still in the process of being created. Now, the
master returns a specific error. This code fixes the table locations lookup code
in AsyncKuduClient for look for that specific error code.

This change isn't tested since the locateTablets codepath is well covered by
existing tests.

Change-Id: I80bf5661aed1ba435800211411b3273874e5bbcf
Reviewed-on: http://gerrit.cloudera.org:8080/3303
Reviewed-by: Jean-Daniel Cryans
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins