tablet: fix sloppiness around atomic mutation lists
Previously we had some TSAN suppressions around our handling of
the Mutation linked lists. This removes them and does the appropriate
Acquire_Loads that we were missing.

In practice I don't think we were hitting any bugs due to this, but
it's hard to know -- maybe this was a cause of some recent crashes?
Either way, fewer suppressions is better.

Change-Id: Idf4e404abd2729660a38841759d3894c4af5eb20
Reviewed-on: http://gerrit.cloudera.org:8080/3410
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>