Fix flakiness of RaftConsensusITest.TestLogIOErrorIsFatal
This test was flaky because it was assuming that the order of entries in the
TServersDetail map matched the ordering of tablet servers in the minicluster's
tablet server vector. This is not always the case, so the assertions waiting
for nodes to crash sometimes timed out (since they were waiting on the wrong
nodes).

Change-Id: If62911f9517b60d82fac97bacd2ac01584be442f
Reviewed-on: http://gerrit.cloudera.org:8080/3493
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins