Fix flakiness in RaftConsensusITest.TestMasterNotifiedOnConfigChange
This test was supposed to wait until the master had committed
a given operation, but in fact was only waiting until all of the
servers had _replicated_ the operation. In TSAN builds where things
ran slower, this caused it to be about 7% flaky.

The fix is to simply wait for the op to be committed, not just replicated.
With the fix, I looped it 100 times and it passed.

Change-Id: Ibb5875e8d4200b0058cf7e7c4ee04771b91d2a24
Reviewed-on: http://gerrit.cloudera.org:8080/3511
Tested-by: Todd Lipcon <todd@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>