ts_recovery-itest: fix flakiness in TestRestartWithPendingCommitFromFailedOp
This test seems to be flaky due to not waiting long enough for the TS to crash.
I tried to repro on dist-test, but the GCE test slaves are fast enough that it
passed 1000/1000. On EC2 apparently things are a little slower so 10 seconds
isn't enough time to guarantee that the injected fault fires.

Change-Id: I60db2acf7642aa6eb2c850032a7dc9639b6ae40e
Reviewed-on: http://gerrit.cloudera.org:8080/3514
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>