master_failover-itest: eliminate some flakiness
Due to the inherent asynchronicity of SIGSTOP and the lack of "exactly once"
semantics [1], it's possible for any DDL operation issued by these tests to
return a result suggesting it had already been performed prior to being
issued. See TestCreateTableSync for a detailed example.

I've never run into this myself, but it has occurred at least once [2] in
the wild.

1. EO semantics are being implemented as part of the replay cache work, but
   to fix this particular problem we'll need some additional logic (i.e.
   we'd need to copy RPC information from e.g. CreateTableRequestPB into the
   replicated WriteRequestPB that corresponds to the created table).
2. http://dist-test.cloudera.org:8080/diagnose?key=d73c47c0-3c15-11e6-8c40-0242ac110001

Change-Id: Ieba6da4d2a4333760022c68783c32dc2689a8a26
Reviewed-on: http://gerrit.cloudera.org:8080/3521
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>