KUDU-1490: Show software version of tablet servers
In the web UI, it could be helpful to know what version of the
software a tablet server is running. This commit addresses this
by adding a field to the initial registration heartbeat of a
tablet. The version is displayed in the Registration column of
the Tablet Servers page of the master web server.

See:
https://raw.githubusercontent.com/anjuwong/kudu/tserver-vers/tserver_vers_on_master_ui.png

Change-Id: I84b32608c9e12a47f422b5047591d4098d187478
Reviewed-on: http://gerrit.cloudera.org:8080/3535
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins