Disable core dumping in integration tests which crash
In TSAN builds, core dumps can take minutes as the virtual address space
is very large. In a few new tests, we injected IO errors and expected
the tserver to crash. In TSAN builds, this crash could take tens of seconds
due to the very large core dumps, and thus result in the test failing.

This is similar to KUDU-942 (and also d3fc252a12f08ba) where we disabled
core dumps for injected crash faults. In these new cases, however, we don't
directly inject a crash, but rather inject an IO error which manifests
itself as a crash.

Change-Id: I2eca1395fa3438861ee1b1ff49978408fed390f4
Reviewed-on: http://gerrit.cloudera.org:8080/3534
Reviewed-by: Mike Percy <mpercy@apache.org>
Tested-by: Kudu Jenkins