[java client] RPCs can get lost in a TabletClient race
We saw hangs after running ITBLL for hours. Turns out that the recent
fixes in TabletClient introduced a new race condition. rpcs_inflight
is being cleaned in cleanup() by copying all the elements from it
and then calling clear(). Even though this is done under a lock, that
lock isn't protecting rpcs_inflight so it's possible to clear() rpcs
that were not copied out.

I haven't been able to recreate this race in unit tests, but it fixed
ITBLL.

Change-Id: Iaff89eb832d0d6f0dede198661856fae1a8585a0
Reviewed-on: http://gerrit.cloudera.org:8080/3541
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>