deltafile-test: fix TestEmptyFileIsAborted for file block manager
This should improve life for the macOS folks.

I thought about explicitly aborting the block (to delete it) in
DeltaFileWriter::Abort(), but reasoned that the impact would be minimal, as
DeltaFileWriters are short-lived to begin with, and as best I can tell, no
code depends on those blocks vanishing immediately.

Change-Id: Icc27ac5555ceb49a010b2b4fbc5123e0c4fdfb5a
Reviewed-on: http://gerrit.cloudera.org:8080/3546
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>